requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pillow==10.2.0
pytesseract==0.3.10
python-multipart==0.0.9
pyzbar==0.1.9
numpy==1.26.4
opencv-python-headless==4.9.0.80
